Trac #33091: fix mislabeled multiline "long time" tests.
There are a few tests in the sage library that span multiple lines and
that were intended to be marked as "long time", but were not due to an
implementation detail. As it stands, it is the FIRST line of a doctest
that needs to be marked "long time", and not the last.

This commit fixes a few doctests whose last (but not first) lines were
tagged, updating them instead to bear the "# long time" tag on each
line. This is overkill, but will save us time in the event that the
parser is ever changed.