src/sage/doctest/forker.py: report verbose doctest success in CPU time
Since we're now warning about tests that take too many cpu-seconds (as
opposed to too many wall-seconds), it makes sense to report the time
taken in cpu-seconds as well. This allows you to better judge how
close a test is to being a "long" test. The output format remains
unchanged, for example,

  ok [1.02 s]

but now the "s" refers to cpu-seconds instead of wall-seconds. A
subsequent commit will disambiguate the two types of "s".

Issue: GH-32981