gap: do not directly dlopen() the gap library
This needs the soname (as in sage.env.GAP_SO) which has issues for
system gap as explained in #33446.

Instead we dlopen() the extension module sage.libs.gap.util which,
having a link time dependency to libgap, will indirectly dlopen() it.

For the record: by the time we run dlopen() the libgap should be already
loaded. The purpose of doing it is to change mode to RTLD_GLOBAL so that
symbols in libgap are placed in the global symbol table. This is
required to compiled load gap packages.

An easy test that this is working ok is:

    sage: libgap.LoadPackage("io")
    true

This requires optional spkg `gap_packages` to be installed.