Trac #32867: don't set a heap limit for the pexpect maxima.
We have some old code and an old test for a memory limit that was
present in ECL 12 years ago, worked around in ticket 6772. The memory
limit is no longer in place, and would not have applied to a non-ECL
maxima from the system anyway. Since the test itself is purposely
wasteful (to ensure that we can use lots of RAM), we take this
opportunity to remove it and speed up the test suite a bit.