Trac #32867: don't require pexpect maxima to live in a "/bin" directory.
This just fixes a test. The problem was noticed on Arch where, due to
some other circumstances, maxima was executed from within /usr/sbin.
But all we know for sure is that maxima will live in PATH. So we
modify the test to look only for "maxima", and not its parent
directory.