Define __init__ and self._frobenius_trace/norm for FiniteDrinfeldModule
I wanted to cache the Frobenius trace and norm, as to compute them only
once. For this, I needed to add

    self._frobenius_norm = None
    self._frobenius_trace = None

in __init__.

However, __init__ did not exist (we simply did not need it), so I
created it.

If there is any better way to set self._frobenius_norm and
self._frobenius_trace at init, please tell me.