Change _get_action_ to action in DrinfeldModule
See https://trac.sagemath.org/ticket/33713 for a discussion on this
topic. Representations are very unusual in this context; this
implementation seems more appropriate.

I added documentation for this method in DrinfeldModule main docstring.

NOTE: The commit is not very clean as I did some cleaning in the
docstring too.