Change base for Drinfeld modules classes
The so-called base of a Drinfeld module (or its category) was a field
(the field of coefficients of the Ore polring). This was fundamentally
misleading, as the good notion for the base is that of a morphism Fq[X]
-> K.

The base is now this morphism.

Method change_ring was deleted.

Doc is updated; tests pass.

Some changes were made to other _repr_ and _latex_ methods.