Fix setuptools failure to build on MSYS
As mentioned in the comment, MSYS changing the default OS separator for
the Windows Python breaks setuptools attempts to install itself, so this
fixes that.

Long term, the right fix for this is probably replacing the currently
deprecated `setup.py install` command with `pip install` in some form,
since the bug seems to occur in setuptools' installation logic (and
using `pip install` would probably use pip for that installation logic,
instead of setuptools).

That being said, this is a really hacky but technically functional fix
that seems to work for now.