Read and write from tmpfile to not block queue
it appears that you cannot send arbitrary amounts of data through a
message queue in Python, see
https://stackoverflow.com/questions/10028809/maximum-size-for-multiprocessing-queue-item.

The queue will block pretty early. So we take load off the queue by
using the local file system for some of the transfers.

We still need to delete the files which is not done in this proof of
concept yet.