Fix deployment of docs to netlify (#34984)
Normally, on github, netlify takes care of the build + deploy fully
automatically. However, our build takes to long so that we would hit the
free build mins relatively quickly. So we continue deploying the docs
via github actions. But PR workflows don't have access to secrets, so we
need to split the upload of the docs in a new workflow, see
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/.

Moreover, the netlify deploy cli doesn't take care of adding a PR
comment and status check linking to the deployed website (netlify does
this automatically if it would build + deploy the page). So we add this
functionality manually.