#26968: add a doctest to catch ecl race in maxima init
We run a new instance of sage in a subprocess to ensure maxima is
not already initialized. We use a temporary MAXIMA_USERDIR so its empty,
and we try to initialize maxima twice in parallel to entice the race.

This temporary dir is placed within `DOT_SAGE` so it is easy
to try different filesystems.

The bug triggers more frequently if `DOT_SAGE` is in a high
latency filesystem (e.g. sshfs on a non-local host).

The next commit introduces a workaround for the bug.