Do not convert lists to FreeModuleElement when constructing FiniteFieldElement_pari_ffelt
When doing intensive polynomial arithmetic with the NTL implementation
the constructor with lists is called a large number of times
and may spend a lot of time constructing the vector_space and
FreeModuleElement objects, instead of using the list directly.

Similar code for number fields also considers lists and vectors
equivalently.