Use NTL factoring functions for factor and squarefree_decomposition
Similarly to polynomials implemented through FLINT,
use NTL factoring API for NTL polynomials (over GF(2), GF(p) for large p
and field extensions).

The previously used PARI implementation is still accessible
through algorithm=pari.

The NTL implementation is made the default: for GF2X polynomials,
it is orders of magnitude faster than PARI implementation,
and for large prime fields, it is noticeably faster.
It is also not subject to the limitation of PARI stack size.

For finite field extensions, the PARI implementation is faster
and remains the default choice.