Reimplement qsieve() using FLINT's qsieve_factor().
The qsieve() function from sage.interfaces.qsieve uses an old,
standalone copy of Bill Hart's quadratic sieve that was moved into its
own package, called FlintQS. This implementation isn't terribly
reliable by modern standards (just look at the docstring), and it
requires maintenance for security issues and to appease newer
compilers.

However, in the intervening years since FlintQS was forked, a
quadratic sieve was added to FLINT itself, as qsieve_factor(). This is
a bit slower than FlintQS in general, but returns the right answer
much more often. Instead of bringing FlintQS up to date, this commit
reimplements sage's qsieve() with FLINT's qsieve_factor().

This entails adding qsieve_factor() to sage.libs.flint, and rewriting
qsieve() to call it and process the output. The implementation is very
similar to factor(n, algorithm="flint").

The output type of the function has necessarily changed, since we now
provide not only the factors but their exponents as well.