src/doc/en/thematic_tutorials: update material relating to qsieve().
The integer factorization tutorial suggests that qsieve() is the best
algorithm, that it is faster than PARI, and that sage's generic
factor() should be rewritten to use qsieve() and ecm.factor(). We have
toned down these claims:

  * PARI has gotten much faster and is competitive with qsieve()
    now that qsieve() is implemented with qsieve_factor() and
    returns fewer wildly incorrect answers.

  * I highly doubt that it's obvious which algorithm will work the
    best before you actually try them, making it very hard to improve
    on what PARI or FLINT are doing in a generic factor().

The (not tested) timings have also been removed so that we can test
the output. Not only are the timings outdated and random, but they
aren't even comparatively accurate anymore (for example, PARI doesn't
take noticeably longer to factor the integer we give it than qsieve
does). In any case, it's nice to check that the examples work.