Don't force ecl lisp with `maxima -l ecl` on command line..
Instead, use `maxima` which should run maxima under the default lisp.
This is consistet with what happens with `sage.interfaces.maxima`.

Note that `maxima_lib` needs maxima built with ecl, but the pexpect
interface to maxima works with any lisp and there are faster builds of
maxima (e.g. the one compiled with sbcl).