Fix doctest so it works with old and new giac.
This `...` corresponds to a warning in the version of giac bundled with
sage, but the warning is removed in current version of giac (tested with
1.9.0-43).

Before this `...` was ok since it was matching giac initialization, but
after the last commit this is no longer the case.

Placing the `...` in the same line as the output makes it optional.