Support linbox 1.7.0 and 1.6.3 at the same time
The constructor `DenseMatrix(field, entries, m, n)` in 1.6.3 was changed
to `DenseMatrix(field, m, n, entries)` in 1.7.0.

We replace use of this constructor by a different way to initialize a
matrix from entries using part of the API that works both in 1.6.3 and
in 1.7.0.

Note that the two versions of the library are still ABI-incompatible so
changing linbox requires recompiling sagelib. To complicate matters,
linbox doesn't change soname so the dynamic linker will let this go and
sagemath will eventually crash.