GAP: fix definition of 'Int'
The GAP Int type, like UInt, has the size of a machine pointer (at least
on the usual desktop system architectures), not a machine int. I.e. on a
typical 64bit OS it is a 64bit integer, not a 32bit integer. This often
makes no difference on little endian systems but can have disastrous
consequences for large values or on big endian systems.

Also fix a copy&paste mistake in the `make_gap_integer` argument name.