gap: use CALL_WITH_STREAM to redirect output to string
This avoids use of deeply internal GAP APIs which may change at any time.
In particular `TypOutputFile` is likely to change again and again.

In contrast, `CALL_WITH_STREAM` is on a much higher level (and even
reachable from the GAP language itself). We have no plans to change it.