Bug fix for isogeny applied to point of infinite order
This is a bug fix for the following bug in ell_curve_isogeny.py:
If phi is an isogeny E --> E' and a is a point of infinite order
then computing a.order() and then phi(a) will throw an exception.
We now check if P._order == +Infinity prior to computing
P._order.gcd(...) to avoid the exception.