src/sage/doctest/control.py: log cysignals crashes to DOT_SAGE
We set CYSIGNALS_CRASH_LOGS="$DOT_SAGE/crash_logs" in sage-env, but
currently, are are overriding that with a temporary directory in the
DocTestController's run_val_gdb() method.

For consistency, it makes more sense to place the doctest crash logs
in the same place as the "regular" crash logs. Doctest crashes are
important too, and if they are worth retaining at all, they are worth
retaining beyond a reboot or ten days (the current systemd
default). On the other hand, if the logs were truly transient, they
could be disabled by setting CYSIGNALS_CRASH_LOGS to an empty value.
Since we are making some effort to save them, they should be
considered more persistent than not.