src/sage/doctest/control.py: log cysignals crashes to the default path
Our doctest controller is overriding the value of CYSIGNALS_CRASH_LOGS
to save any crash logs arising from a doctest in a tmp_dir(). At some
point in the past, these "temporary" logs would have been somewhat
permanent: the tmp_dir() function did not clean up after itself, and
there were no tmpfiles.d entries to prune /tmp before a reboot. But
both of these circumstances have changed, meaning that the crash logs
now lie in the convex hull of very and extremely temporary.

For consistency, it makes more sense to place the doctest crash logs
in the same place as the "regular" crash logs. Doctest crashes can be
important too, and if they are worth retaining at all, they are worth
retaining beyond the end of the sage session, or a reboot, or ten
days, etc. They should also be a rare occurrence. Moreover if the logs
were never meant to be kept, i.e. if they were truly meant to be
temporary, that could have been achieved with CYSIGNALS_CRASH_LOGS="".
It was not; so in that sense, using the default location restores the
intended behavior.

This commit removes the override, causing the crash logs to be saved
to whatever the current value of CYSIGNALS_CRASH_LOGS is, or to the
current working directory as a fallback. (These are cysignals
implementation details.) Note that the sage-env script sets
CYSIGNALS_CRASH_LOGS to $DOT_SAGE/crash_logs, so for typical sage
users these logs will be saved there.