build/bin/sage-get-system-packages: optionally skip python packages
Packages calling SAGE_PYTHON_PACKAGE_CHECK in spkg-configure.m4 are
python packages whose detection is enabled by passing the
--enable-system-site-packages flag to ./configure. But when the flag
is NOT passed, we don't want to list the corresponding packages in the
output of sage-get-system-packages. This commit checks for the
presence of such an spkg-configure.m4 file and filters those packages
according to the ENABLE_SYSTEM_SITE_PACKAGES variable.