src/sage/tests: fix another "Computational Math..." doctest
One doctest for the Computational Math book fails for me,

  Failed example:
    limit(f(t * cos(theta), t * sin(theta)) / t, t=0)
  Expected:
    cos(theta)^2/sin(theta)
  Got:
    -1/2*(sin(3*theta) + sin(theta))/(cos(2*theta) - 1)

These results are equivalent, and the ugly one simplifies to the
pretty one. We add a full_simplify() to the test to support both
outputs. The discrepancy is most likely due to the version of Giac
installed on the system.