src/sage/doctest/forker.py: report both cpu/wall times for long tests
We need to report the CPU time for tests that run afoul of
--warn-long, because --warn-long is now specified in CPU time. But, we
are still reporting the wall time in several places, such as the
per-file summary. So to keep things as consistent as possible, we
should report the wall time for --warn-long failures as well. This
changes the output of a too-long test from e.g.

  Test ran for 1.23 cpu seconds, check ran for 2.34 cpu seconds

to,

  Test ran for 1.23s cpu, 2.50s wall
  Check ran for 2.34s cpu, 3.12s wall

Issue: GH-32981