src/sage/doctest: specify walltime in per-file doctest report
Now that cpu seconds are used for --warn-long, we should disambiguate
wherever wall seconds are still used. One such place is in the doctest
runner's per-file report, which outputs (for example)

  [14 tests, 0.26 s]

after testing a file with 14 tests that took 0.26 wall seconds. This
commit changes that output to

    [14 tests, 0.26s wall]

which is explicit about the type of measurement used.

Issue: GH-32981