src/sage/doctest: use CPU time for --warn-long
Our --warn-long flag used during testing currently takes "wall time"
as its argument. This is somewhat nice from a user perspective, since
a normal user is quite likely to know what what wall time is, but it's
also a bit nonsensical given the purpose of the flag. The flag's
intent is to warn developers about tests that take too long, so
whether or not it triggers should not depend on the developer's
machine, or what else he's doing while the tests are running -- both
of which do affect the wall time.

CPU time provides a more reliable measurement. (In fact, when no value
for --wall-time was given, we were already using a heuristic to mimic
CPU time.) This commit changes the --warn-long flag to use CPU time
instead of wall time, obviating the need for the heuristic.

Issue: GH-32981