src/sage/doctest/parsing.py: ignore a transient pythran warning
Some numpy.random signatures changed in numpy-1.25.x, and pythran
normally wants to warn about those. It's not clear, however, whether
or not the numpy changes will be reverted. Until we know,
pythran-0.14.0 has a special case to ignore the warnings for
numpy.random. Here we ignore those warnings in the doctest parser as
well, allowing older versions of pythran to be used with numpy-1.25.x.