src/sage/tests/books: tweak a "Computational Math..." doctest
One of the tests for the "Computational Math..." book is plotting
lines on a vector field by calling scipy.integrate.odeint repeatedly.
This test runs,

  y = integrate.odeint(f, k, t)
  p += line(zip(t, flatten(y)))

in an attempt to flatten y, which is returned as a numpy array of
1-dimensional numpy arrays. However, flatten() isn't smart enough to
handle that: it returns a list of 1-dimensional numpy arrays rather
than a list of floating point numbers. And numpy-1.25.x now warns
about casting 1-dimensional arrays to float deep in our plotting
infrastructure.

Instead, we now call y.tolist() before passing it to flatten(), and
this does the right thing.