src/sage/doctest/util.py: refactor exception handling
Instead of handling all of the possible cputime exceptions in the
Timer class's _quick_cputime(), we move it into _pid_cpu_seconds(),
closer to where the errors actually occur. This allows us to re-raise
an OSError in all exceptional cases, making them easier to handle
higher up. In the process some of the exceptions' causes have been
documented.