build/pkgs/rpy2: use SAGE_PYTHON_PACKAGE_CHECK in spkg-configure.m4
Our system package lists are generated with a special case that looks
for SAGE_PYTHON_PACKAGE_CHECK when spkg-configure.m4 is present. The
rpy2 package already had an spkg-configure.m4 relating to --disable-r,
but its spkg-configure.m4 did not use SAGE_PYTHON_PACKAGE_CHECK. As a
result, it was erroneously included in the package list.

We now use SAGE_PYTHON_PACKAGE_CHECK for the usual "check" phase of
SAGE_SPKG_CONFIGURE, which both allows the user to use rpy2 from the
system and fixes the package list.