The mod 2 cohomology ring of a simplicial complex/set as a module over the Steenrod algebra.
Create a new class, CohomologyRing_mod2, for this, because these
Steenrod operations on cohomology are only implemented at the prime
2. Use the new class automatically when users create
`K.cohomology_ring(GF(2))` for a complex `K`. Move the existing "Sq"
method to the new class, and create new methods to make the module
action transparent.

Also create a method steenrod_module_map in case (a) someone wants
access to these matrices and (b) for future use, and in particular for
computing the Steenrod algebra action on homology.