src/sage/misc/latex.py: fix view()
Rewriting the view() function to use python's tempfile module instead
of sage's own tmp_dir() accidentally broke this function, because the
viewer program needs the file to be slightly less temporary than we
made it. The viewer would launch, but then view() would immediately
return, causing the file that the viewer was about to look for to be
deleted.

To fix it, we now launch the viewer program synchronously within a
helper function, but launch that helper function asynchronously in a
new thread. This allows us to clean up the temporary directory only
after the subprocess has completed, but still lets view() return
immediately.

Closes: https://github.com/sagemath/sage/issues/36526