Support networkx 3.2
In networkx 3.2 the output for cycle_basis() has changed.

After a discussion in https://github.com/sagemath/sage/issues/36486 it
seems that checking that the output is correct would not be easy. In the
spirit of supporting networkx 3.2 at the same time as networkx 3.1, the
easiest way is to mark these five tests as random output.