build/pkgs/pari: require pari-2.5.14, drop tests for patched v2.5.13
We currently check for either pari-2.5.14, or an unofficial (patched)
version of pari-2.5.13 that is missing two bugs present in the
upstream release. However, one of those bug-checks fails on macOS,
making it impossible to use the system pari there.

pari-2.5.14 has been out for a few months. Any distro fast enough to
have backported patches to pari-2.5.13 will likely have pari-2.5.14 by
now. On that assumption, we simply require pari-2.5.14 to fix the
macOS issue, and hope it does not affect anyone still using a patched
pari-2.5.13.