build/pkgs/gap{,_packages}: make most GAP packages optional
There's a long list of "recommended" packages that we install for GAP
because, otherwise, GAP complains about them being missing when you
start it. This can be avoided with the -A flag, however, that we are
now passing to both the CLI gap and libgap. These additional packages
are therefore no longer needed to silence the warnings.

Here we move all of these recommended packages from the gap SPKG to
the gap_packages SPKG.