src/doc/en/prep/Quickstarts/Abstract-Algebra.rst: don't list subgroups
The doctest for the normal_subgroups() of DihedralGroup(8) cannot
simply list them all, because it turns out that in GAP the
NormalSubgroups() function has some freedom to choose
representatives. As a result, verbatim output is not guaranteed.

We are already sorting the subgroups as part of an earlier attempt to
solve this problem, so the two trivial subgroups should live at the
beginning and end of the list. As such, we use "..." to match whatever
is output for the intermediate groups. We also add a new test for the
number of normal subgroups.

This fixes a test failure on Conda.