src/sage/groups/matrix_gps/finitely_generated_gap.py: fix a GAP doctest
Our doctest for as_permutation_group() checks that algorithm="smaller"
produces a group that is different from the one obtained with
algorithm=None. But this need not be the case (and sometimes it is
not): the SmallerDegreePermutationRepresentation() function in GAP is
doing its best, but it "might involve the use of random elements and
the permutation representation (or even the degree of the
representation) is not guaranteed to be the same for different calls."

Bottom line, we can't count on the "smaller" group actually being
smaller than, or even different from, what we would get without
algorithm="smaller". So we drop that test, and compare the degrees of
the two groups instead, allowing for equality.

This fixes a failing test on Gentoo and Conda.