build/pkgs/backcall: skip if the system ipython is used
ipython is the only package that depends on backcall, so backcall can
be skipped if ipython comes from the system. This micro-optimization
is not so micro now that ipython has dropped backcall upstream:

  https://github.com/ipython/ipython/commit/1c2687ea5

As a result, systems with newer ipythons may not have backcall
installed. We don't want to install a pointless SPKG on those systems.