src/sage/interfaces/gap.py: don't load all packages on workspace reset
Resetting the GAP workspace already initializes a new Gap instance and
loads our default collection of packages. There's no need to load
anything else in gap_reset_workspace().

The disabling of the color prompt is now handled in both sage.gaprc
and sage.g (both GAP interfaces) so we remove that as well.

Fixes: GH-31761