src/sage/ext_data/gap/sage.g: disable color prompt
This is currently done in gap_reset_workspace(), but I think it's less
surprising to find the GAP initialization in the GAP initialization
file. Here we add it to sage.g (for the pexpect interface) so that it
can later be removed from gap_reset_workspace(). It has already been
added to sage.gaprc (for the library interface) in an earlier commit.