src/sage/tests/gap_packages.py: remove
These tests for the installed GAP packages are problematic in a world
where Sage can use GAP from the system. Particularly, loading _all_
installed GAP packages is not a safe operation when there might be an
enormous number of them.

Moreover, with the system GAP, it's not really Sage's reponsibility to
ensure that the installed packages load. It _would_ be nice to know
that any extra packages from the gap_packages SPKG are installed
correctly; however, that type of test would be more appropriate in
that SPKG's spkg-check phase.